Added a new externalalg product to stitch geo imagery - but this works off of temporary data files, so it probably won't be terribly useful for our current realtime processing, that is temporary imagery based.
This basically requires calling driver with
-s global -p Geo-Stitched --sectorfile=$GEOIPS/geoips/sectorfiles/test/Global.xml --write_sectored_datafile
on the seviri, abi, and ahi files you want included in the stitched products. Since the geo readers
actually register the data on read, the sectored_datafile output is actually
pre-registered.

Then the geo_stitched.py algorithm will go find the appropriate registered data files in:
$GEOIPS_OUTDIRS/preprocessed/sectored/seviri_None/Global_20180722.230000-20180722.231500_meteoEU_seviri_001_2338966969219643690.h5
(That path is specified in $GEOIPS/geoips/utils/plugin_paths.py, and written out in driver.py after the data has been read and sectored)

This is not optimized or error checked at all, but perhaps is a starting point for
intermediate data file products of some sort.