1. Only try to delattr when doing data merges if they exist  (_image and _registered_data) 2. Use correct variable name if otherimg.source_name != self.datafile.source_name (previously just used self.datafile.source_name variable names for both datasets - which doesn't work if you're trying to merge, say, himawari and meteosat)
Updated to allow for different variables names when doing data merges of different sources.
NOTE this works with single channel products, but the order may be screwed up when doing products requiring multiple channels. We're not doing that right now, so if we do in the future pay attention...

We are basically grabbing the list of required variables from the "other" source, then popping them off as we loop through the variables in the current "self" dataset.
If the order happens to match up, they will be merged appropriately, then used in the multi-channel algorithm.
If the order doesn't match up, random variables will be merged together from the different sources, and you could get interesting results....

Note I did try to make this backwards compatible, so it will NOT fail for existing processing on a single source.
We only try to use a different variable name if self.datafile.source_name != otherimg.source_name.